# aws-lc-sys: only build libssl if needed