--- a/aws-lc-sys/CMakeLists.txt
+++ b/aws-lc-sys/CMakeLists.txt
@@ -10,37 +10,49 @@ enable_language(C)
 
 add_subdirectory(aws-lc aws-lc EXCLUDE_FROM_ALL)
 
-add_definitions(-DAWS_LC_RUST_INCLUDE_SSL)
+if (BUILD_LIBSSL)
+    add_definitions(-DAWS_LC_RUST_INCLUDE_SSL)
+endif()
 add_library(rust_wrapper STATIC rust_wrapper.c)
 target_include_directories(rust_wrapper PRIVATE include)
-target_link_libraries(rust_wrapper PUBLIC crypto ssl)
+
+if (BUILD_LIBSSL)
+    target_link_libraries(rust_wrapper PUBLIC crypto ssl)
+else()
+    target_link_libraries(rust_wrapper PUBLIC crypto)
+endif()
 
 set(FINAL_ARTIFACTS_DIRECTORY ${CMAKE_BINARY_DIR}/artifacts)
 
 # Places the compiled library files at the root of the binary directory so we can have a consistent
 # location to find the artifacts cross-platform.
-set_target_properties(rust_wrapper crypto ssl
+set_target_properties(rust_wrapper crypto
         PROPERTIES
         ARCHIVE_OUTPUT_DIRECTORY ${FINAL_ARTIFACTS_DIRECTORY}
         LIBRARY_OUTPUT_DIRECTORY ${FINAL_ARTIFACTS_DIRECTORY})
+
+if (BUILD_LIBSSL)
+    set_target_properties(ssl
+            PROPERTIES
+            ARCHIVE_OUTPUT_DIRECTORY ${FINAL_ARTIFACTS_DIRECTORY}
+            LIBRARY_OUTPUT_DIRECTORY ${FINAL_ARTIFACTS_DIRECTORY})
+endif()
 
 # Based on https://stackoverflow.com/a/7750816 as some generators, like MSVC, will try to prefix the output directory
 # which is not needed in our case.
 foreach (OUT_NAME ${CMAKE_CONFIGURATION_TYPES})
     string(TOUPPER ${OUT_NAME} OUT_NAME)
-    set_target_properties(rust_wrapper crypto ssl
+    set_target_properties(rust_wrapper crypto
             PROPERTIES
             ARCHIVE_OUTPUT_DIRECTORY_${OUT_NAME} ${FINAL_ARTIFACTS_DIRECTORY}
             LIBRARY_OUTPUT_DIRECTORY_${OUT_NAME} ${FINAL_ARTIFACTS_DIRECTORY})
+    if (BUILD_LIBSSL)
+        set_target_properties(ssl
+                PROPERTIES
+                ARCHIVE_OUTPUT_DIRECTORY_${OUT_NAME} ${FINAL_ARTIFACTS_DIRECTORY}
+                LIBRARY_OUTPUT_DIRECTORY_${OUT_NAME} ${FINAL_ARTIFACTS_DIRECTORY})
+    endif()
 endforeach ()
-
-string(TOUPPER $<CONFIG> UPPER_CONFIG_NAME)
-
-set_property(TARGET rust_wrapper crypto ssl
-        PROPERTY ARCHIVE_OUTPUT_DIRECTORY_${UPPER_CONFIG_NAME})
-
-set_property(TARGET rust_wrapper crypto ssl
-        PROPERTY LIBRARY_OUTPUT_DIRECTORY_${UPPER_CONFIG_NAME})
 
 if (BORINGSSL_PREFIX)
     if (MSVC)
@@ -48,5 +60,8 @@ if (BORINGSSL_PREFIX)
     else()
         set(TARGET_PREFIX "lib${BORINGSSL_PREFIX}")
     endif()
-    set_target_properties(rust_wrapper crypto ssl PROPERTIES PREFIX ${TARGET_PREFIX})
+    set_target_properties(rust_wrapper crypto PROPERTIES PREFIX ${TARGET_PREFIX})
+    if (BUILD_LIBSSL)
+        set_target_properties(ssl PROPERTIES PREFIX ${TARGET_PREFIX})
+    endif()
 endif ()