### Issues:
N/A

### Description of changes: 
This change causes the build to only create `libssl` when the "ssl" feature is enabled. This improves overall build latency.

Before:
```
❯ time cargo build 
   Compiling cc v1.0.79
   Compiling libc v0.2.142
   Compiling paste v1.0.12
   Compiling dunce v1.0.4
   Compiling cmake v0.1.50
   Compiling aws-lc-sys v0.6.1 (/home/justsmth/repos/aws-lc-rs/aws-lc-sys)
    Finished dev [unoptimized + debuginfo] target(s) in 11.36s
cargo build  54.37s user 8.81s system 554% cpu 11.395 total
```

After:
```
❯ time cargo build
   Compiling cc v1.0.79
   Compiling libc v0.2.142
   Compiling paste v1.0.12
   Compiling dunce v1.0.4
   Compiling cmake v0.1.50
   Compiling aws-lc-sys v0.6.1 (/home/justsmth/repos/aws-lc-rs/aws-lc-sys)
    Finished dev [unoptimized + debuginfo] target(s) in 6.41s
cargo build  24.37s user 5.40s system 461% cpu 6.449 total
```

### Call-outs:
N/A

### Testing:
I've run the crate generation scripts.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and the ISC license.